% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_fac}
\alias{compute_smd_fac}
\title{SMD computation for factor variables}
\usage{
compute_smd_fac(x1, x2, wts1 = NULL, wts2 = NULL, digits = 3)
}
\arguments{
\item{x2}{}
}
\description{
SMD computation for factor variables
}
