% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_num}
\alias{compute_smd_num}
\title{SMD computation for numeric vectors}
\usage{
compute_smd_num(x1, x2, wts1 = NULL, wts2 = NULL, abs = TRUE, digits = 3)
}
\arguments{
\item{abs}{}
}
\description{
SMD computation for numeric vectors
}
