% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd_num_nn}
\alias{smd_num_nn}
\title{SMD for skewed/non-normal numeric variables}
\usage{
smd_num_nn(x, grp, ...)
}
\arguments{
\item{x}{Numeric vector of data values.}

\item{grp}{A factor of same length as `x` specifying groups.
If not a factor, the vector will be coerced.}

\item{...}{Additional arguments passed to `smd_num`.}
}
\value{
A named numeric vector with one SMD per group pair.
}
\description{
Computes SMD using ranks of the data values rather than the values themselves.
}
