% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ_fac}
\alias{summ_fac}
\title{Summarize categorical variables}
\usage{
summ_fac(x, wts = NULL, digits = 0)
}
\arguments{
\item{x}{A factor vector.}

\item{wts}{A non-negative vector of weights.}

\item{digits}{Number of decimal places to use.}
}
\value{
A named character vector with "n (%)" for each level, or just "%" if weighted.
}
\description{
Summarize categorical variables
}
