% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ_num}
\alias{summ_num}
\title{Summarize numeric variables}
\usage{
summ_num(x, wts = NULL, digits = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{wts}{A numeric vector of weights (optional).}

\item{digits}{Number of digits to round to (default is 2).}
}
\value{
Returns (weighted) arithmetic mean of values in x.
}
\description{
Summarize numeric variables
}
