% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1-package.R
\docType{package}
\name{tab1-package}
\alias{tab1-package}
\title{tab1: Create Publication Quality Table 1}
\description{
Create publication quality Table 1.
}
\author{
\strong{Maintainer}: Karan Shah \email{karankshah95@gmail.com}

}
\keyword{internal}
