% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Construct table 1}
\usage{
tab1(
  data,
  nonnormal = NULL,
  opts_num = NULL,
  opts_num_nn = NULL,
  opts_cat = NULL
)
}
\arguments{
\item{nonnormal}{}
}
\description{
Construct table 1
}
