% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_num}
\alias{tab1_num}
\title{Table 1 for numeric variables}
\usage{
tab1_num(data, wts = NULL, digits = 2, ...)
}
\arguments{
\item{data}{A data frame}

\item{wts}{A non-negative numeric vector of weights.}

\item{digits}{Number of digits to use for summary stats.}

\item{...}{Other arguments passed to summ_num.}
}
\value{
A data frame.
}
\description{
Table 1 for numeric variables
}
