% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{compute_smd_num}
\alias{compute_smd_num}
\title{SMD computation for numeric vectors (supports only two groups)}
\usage{
compute_smd_num(x1, x2, abs)
}
\arguments{
\item{abs}{}
}
\description{
SMD computation for numeric vectors (supports only two groups)
}
