% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_tab1.R
\name{pub_tab1}
\alias{pub_tab1}
\title{Create publication quality table 1 from tab1 object}
\usage{
pub_tab1(
  tab1,
  lbl_nm = "mean (SD)",
  lbl_nn = "median [Q1, Q3]",
  lbl_fac = "n (\%)"
)
}
\arguments{
\item{tab1}{Data frame output from `tab1` function.}

\item{lbl_nm}{Summary statistic label for numeric variables.}

\item{lbl_nn}{Summary statistic label for non-numeric variables.}

\item{lbl_fac}{Summary statistic label for factor variables}
}
\value{
A flextable object.
}
\description{
Create publication quality table 1 from tab1 object
}
\examples{
tab1(iris, "Species") |>
pub_tab1()
}
