% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd_num_nn}
\alias{smd_num_nn}
\title{SMD for skewed/non-normal numeric variables}
\usage{
smd_num_nn(x, grp, ...)
}
\arguments{
\item{x}{Numeric vector of data values}

\item{grp}{Vector of same length as `x` specifying grouping for for
data values}

\item{...}{Additional arguments passed to `smd_num`.}
}
\description{
SMD for skewed/non-normal numeric variables
}
