% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ_fac}
\alias{summ_fac}
\title{Summarize categorical variables}
\usage{
summ_fac(x, wts = NULL, digits = 0)
}
\arguments{
\item{x}{A factor vector.}

\item{wts}{A non-negative vector of weights.}

\item{digits}{Number of decimal places to use}
}
\value{
(Weighted) proportions for values in `x`.
}
\description{
Summarize categorical variables
}
