% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ_num_nn}
\alias{summ_num_nn}
\title{Summarize non-normal numeric variables}
\usage{
summ_num_nn(x, wts = NULL, digits = 0, delim = ",")
}
\arguments{
\item{delim}{}
}
\description{
Summarize non-normal numeric variables
}
