% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Create Table 1}
\usage{
tab1(
  data,
  grp,
  vars = NULL,
  nonnormal = NULL,
  lbl = NULL,
  opts_summ = list(num_digits = 2, numnn_digits = 0, fac_digits = 1)
)
}
\arguments{
\item{lbl}{A named list of the form `list(old_name = new_name)` to rename the
variables in tab1. Must be of the same length as `var` col in tab1.}

\item{opts_summ}{}
}
\description{
Create Table 1
}
