% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Create Table 1}
\usage{
tab1(
  data,
  grp,
  wts = NULL,
  vars = NULL,
  nonnormal = NULL,
  lbl = NULL,
  opts_summ = list(num_digits = 2, numnn_digits = 0, fac_digits = 1)
)
}
\arguments{
\item{data}{A data frame.}

\item{grp}{A character variable specifying the grouping variable.}

\item{vars}{A character vector specifying the variables to summarize.}

\item{nonnormal}{A character vector specifying which variables to treat as
skewed/nonnormal.}

\item{lbl}{A named list of the form `list(old_name = new_name)` to rename the
variables in tab1. Must be of the same length as `var` col in tab1 (optional)}

\item{opts_summ}{A list specifying additional options for summarizing.}
}
\description{
Create Table 1
}
