% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Create Table 1}
\usage{
tab1(
  data,
  grp,
  wts = NULL,
  vars = NULL,
  nonnormal = NULL,
  lbl = NULL,
  opts_summ = list(num_digits = 2, numnn_digits = 0, fac_digits = 1),
  opts_smd = list(denom = "unweighted", abs = TRUE)
)
}
\arguments{
\item{data}{A data frame.}

\item{grp}{A character variable specifying the grouping variable.}

\item{wts}{A (non-negative) numeric vector of weights (optional). See details.}

\item{vars}{A character vector specifying the variables to summarize.}

\item{nonnormal}{A character vector specifying variables to be treated as
skewed/nonnormal.}

\item{lbl}{A named list of the form `list(old_name = new_name)` to rename the
variables in tab1. Must be of the same length as the `vars` parameter (optional).}

\item{opts_summ}{A list specifying additional options for summarizing variables.
Currently supports the following options:
 - num_digits
 - numnn_digits
 - fac_digits}

\item{opts_smd}{A list specifying options for SMD calculations. Supports the
following options:
 - denom: A character vector specifying how to handle denominators when
 calculating SMD. See details.
 - abs: Logical specifying whether to return absolute SMD. Default is TRUE.}
}
\value{
A data frame with summary statistics and pairwise SMDs.
}
\description{
Creates a table with (weighted) summary statistics and pairwise SMDs from a given
data frame.
}
\details{
The summary statistics returned are mean (SD), median [Q1, Q3]
and n (%) for numeric, non-normal numeric and factor variables respectively.
If using weights, the function will only report proportions for the variables
since counts are not meaningful due to weights not being normalized. Furthermore,
the sample size reported will be the effective sample size for each group calculated as
sum(w)^2/sum(w^2), where w is the vector of weights.
For weighted data, the standardized mean difference (SMD) can be calculated using
the weighted or unweighted pooled variance in the denominator. However,
using the weighted pooled variance may result in a paradoxical situation where
the mean difference decreases but the SMD increases. It is thus recommended
that the unweighted pooled variance be used to ensure that the same
standardization factor is used both before and after weighting.
}
\examples{
# Run with defaults
tab1(iris, grp = "Species")
# Available options
tab1(iris, grp = "Species", nonnormal = "Sepal.Width",
opts_summ = list(num_digits = 3, numnn_digits = 2),
opts_smd = list(abs = FALSE))
# Weighted data
set.seed(123)
w <- runif(nrow(iris))
tab1(iris, "Species", wts = w)
# Compare weighted vs unweighted denominator for SMD
tab1(iris, "Species", wts = w, opts_smd = list(denom = "weighted"))
}
