% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1}
\alias{tab1}
\title{Construct table 1}
\usage{
tab1(
  data,
  nonnormal = NULL,
  digits_num = 1,
  digits_num_nn = 0,
  digits_factor = 0
)
}
\arguments{
\item{nonnormal}{}
}
\description{
Construct table 1
}
