% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_num_nn}
\alias{tab1_num_nn}
\title{Table 1 for skewed/non-normal numeric variables}
\usage{
tab1_num_nn(data, opts = NULL)
}
\arguments{
\item{opts}{}
}
\description{
Table 1 for skewed/non-normal numeric variables
}
