% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_ug}
\alias{tab1_ug}
\title{Construct table 1 (no groups)}
\usage{
tab1_ug(
  data,
  wts = NULL,
  nonnormal = NULL,
  num_digits = 2,
  numnn_digits = 0,
  fac_digits = 1
)
}
\arguments{
\item{data}{A data frame.}

\item{nonnormal}{Character vector specifying non-normal variables.}

\item{num_digits}{Number of digits for numeric variables.}

\item{fac_digits}{Number of digits for summaries of factor variables.}

\item{num_nn_digits}{Number of digits for sumamry of non-normal numeric variables.}
}
\description{
This function is the workhorse of the `tab1` function. It is not intended
to be called directly.
}
