% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_ug}
\alias{tab1_ug}
\title{Construct table 1 (without groups)}
\usage{
tab1_ug(
  data,
  nonnormal = NULL,
  opts_num = NULL,
  opts_num_nn = NULL,
  opts_cat = NULL
)
}
\arguments{
\item{data}{data}

\item{nonnormal}{data}

\item{xyz}{}
}
\description{
This function is the workhorse of the `tab1` function but not intended
to be called directly.
}
